Windows_ver/requirements.txt
PyQt5
beautifulsoup4
requests
websocket-client
lxml